% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metric_plot.R
\name{combine_metric_plot}
\alias{combine_metric_plot}
\title{Plot combined model performance metrics}
\usage{
combine_metric_plot(
  cmetrics,
  type = c("bias_barplot", "MAIPE_barplot", "IF20_plot", "IF30_plot")
)
}
\arguments{
\item{cmetrics}{Named list returned by \code{\link[=combine_metrics]{combine_metrics()}}, containing
combined model performance metrics.}

\item{type}{Character string specifying the plot type to generate.
Available options are:
\itemize{
\item \code{"bias_barplot"}: bar plot of relative bias
(\code{rBIAS}) with confidence intervals.

\item \code{"MAIPE_barplot"}: bar plot of mean absolute individual
prediction error (\code{MAIPE}).

\item \code{"IF20_plot"}: bar plot of \code{IF20} values with
reference threshold.

\item \code{"IF30_plot"}: bar plot of \code{IF30} values with
reference threshold.
}}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates comparative visualizations of evaluation metrics across multiple
externally evaluated models combined with \code{\link[=combine_metrics]{combine_metrics()}}.
}
\details{
This function is intended for model comparison workflows, allowing visual
inspection of predictive performance across evaluation occasions..

Metrics are displayed separately for each evaluation occasion
(\code{OCC}) using faceted plots, enabling direct visual comparison
between candidate models.
}
\examples{

# Fake  metrics
generate_fake_metrics <- function(n_occasions = 3) {
data.frame(
 OCC = rep(1:n_occasions),  # Simula varias ocasiones
 rBIAS = stats::rnorm(n_occasions, mean = 0, sd = 10),
 rBIAS_lower = stats::rnorm(n_occasions, mean = -5, sd = 5),
 rBIAS_upper = stats::rnorm(n_occasions, mean = 5, sd = 5),
 MAIPE = stats::runif(n_occasions, min = 10, max = 50),
 IF20 = stats::runif(n_occasions, min = 20, max = 80),
 IF30 = stats::runif(n_occasions, min = 30, max = 90)
)
}

simulation1 <- list(metrics_means = generate_fake_metrics())
simulation2 <- list(metrics_means = generate_fake_metrics())
# List of models
models_list <- list(
list(model_name = "Test_Model1", metrics_list = simulation1),
list(model_name = "Test_Model2", metrics_list = simulation2)
)
combined_results <- combine_metrics(models_list)

# plot results
combine_metric_plot(combined_results, type = 'bias_barplot')

}
\seealso{
\code{\link[=combine_metrics]{combine_metrics()}}, \code{\link[=plot.EvalPPK]{plot.EvalPPK()}}
}
