% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_path.R
\name{get_model_path}
\alias{get_model_path}
\title{Get path to PK model file}
\usage{
get_model_path(name, ext = c("cpp", "R"))
}
\arguments{
\item{name}{Model name}

\item{ext}{File extension ('cpp' or 'R')}
}
\description{
Get path to PK model file
}
