% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List available built-in PK models}
\usage{
list_models(drug = NULL, author = NULL)
}
\arguments{
\item{drug}{Optional drug name filter.}

\item{author}{Optional author name filter.}
}
\value{
A data frame with available models metadata.
}
\description{
Displays the curated pharmacokinetic models available within the package.
}
\examples{
list_models()
list_models(drug = "Tacrolimus")
list_models(author = "Han")
}
