% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model_code.R
\name{load_model_code}
\alias{load_model_code}
\title{Load built-in model code}
\usage{
load_model_code(model_id)
}
\arguments{
\item{model_id}{Character string with the model identifier.}
}
\value{
A character string containing the model code.

#' @details
The returned object is a character string compatible with
\code{mrgsolve::mcode()} and can be directly used in package functions
that accept model code as character input (e.g.
\code{exeval_ppk()} or \code{run_MAP_estimations()}).
}
\description{
Retrieves population pharmacokinetic model code from the internal
curated model library included in the package.
}
\examples{
mod <- load_model_code("TAC_Han2011")

# Run evaluation
# res <- exeval_ppk(
#   model = mod,
#   model_name = "TAC_Han2011",
#   data = my_data
# )
}
