% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exeval_models.R
\docType{data}
\name{exeval_models}
\alias{exeval_models}
\title{Built-in pharmacokinetic models}
\format{
A data frame with 6 variables:
\describe{
\item{Label}{Unique model identifier}
\item{Drug}{Drug name}
\item{Author}{First author}
\item{Year}{Publication year}
\item{Ref}{Name of Article}
\item{Model_code}{Model code as character string}
}
}
\usage{
exeval_models
}
\description{
Curated population pharmacokinetic models available within the package.
}
\keyword{datasets}
