% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exeval_ppk.R
\name{exeval_ppk}
\alias{exeval_ppk}
\title{External evaluation for Population Pharmacokinetic (popPK) Models}
\usage{
exeval_ppk(
  model,
  data,
  model_name = NULL,
  drug_name = NULL,
  tool = "mapbayr",
  check_compile = TRUE,
  num_occ = NULL,
  num_ids = NULL,
  sampling = TRUE,
  occ_ref = NULL,
  evaluation_type = c("sequential_updating", "stepwise_updating",
    "sequential_reference_updating", "backward_reference_updating"),
  method = c("L-BFGS-B", "newuoa"),
  assessment = c("a_priori", "Bayesian_forecasting", "Complete"),
  verbose = FALSE
)
}
\arguments{
\item{model}{Either:
\itemize{
  \item A character string containing the pharmacokinetic model code written
        in \code{mrgsolve} format.
  \item A pre-compiled \code{mrgmod} object (S3 class from \code{mrgsolve}).
}
If a character string is provided, \code{model_name} must also be specified.}

\item{data}{Data frame. Contains the input data for the estimations, including columns like ID, TIME, and OCC.}

\item{model_name}{Character string. Name of the model.
Required only if \code{model} is provided as character model code.#'}

\item{drug_name}{Character string. Used only for reporting purposes.}

\item{tool}{Character string. Specifies the tool to use for estimation. Currently "mapbayr" is the only option.}

\item{check_compile}{Logical. If `TRUE`, checks if the model compiles correctly in `mapbayr`.}

\item{num_occ}{Integer. Number of occasions (OCC) to include in the analysis. If `NULL`, all unique occasions in `data` are used.}

\item{num_ids}{Integer. Number of unique IDs to include in the analysis. If `NULL`, all IDs are included.}

\item{sampling}{Logical. If `TRUE`, randomly samples the specified number of IDs from the data.}

\item{occ_ref}{Integer. Reference occasion for evaluation types that require a reference. Must be consistent with `evaluation_type`.}

\item{evaluation_type}{Character string. Specifies the evaluation type. Options are:
\itemize{
  \item "sequential_updating": Uses all data up to each occasion.
  \item "stepwise_updating": Uses only the most recent occasion.
  \item "sequential_reference_updating": Uses all data up to a reference occasion.
  \item "backward_reference_updating": Uses the most recent occasion relative to a reference.
}}

\item{method}{Character vector. Specifies optimization methods for `mapbayr`. Options are "L-BFGS-B" or "newuoa".}

\item{assessment}{Character string. Specifies the type of prediction to perform. Options are:
\itemize{
  \item "a_priori": Simulates concentrations using the population model without individual data.
  \item "Bayesian_Forecasting": Simulates concentrations using individual parameter estimates (posterior mode).
  \item "Complete": Performs both a priori and Bayesian forecasting simulations.
}}

\item{verbose}{Logical. If TRUE, messages are printed during execution.
If FALSE (default), errors are stored as warnings accessible with `warnings()`}
}
\value{
A list containing:
\describe{
  \item{metrics}{Evaluation metrics}
  \item{estimations}{MAP estimation results for each subset of the data.}
  \item{simulation}{A list of simulation results for each occasion and individual.}
  \item{updates}{A list containing posterior estimations (`a.posteriori`) for each occasion.}
}
}
\description{
Performs external evaluation of popPK models by conducting MAP estimations and individual simulations
for each occasion using different evaluation strategies (see evaluation_type)
}
\examples{
\dontrun{
data("tacrolimus_pk1_kidney", package = "preDose")
data("model_tacHAN2011", package = "preDose")

dd <- tacrolimus_pk1_kidney |> subset(ID < 6)
mm <- get_model_code('Han_etal_2011')

res <- exeval_ppk(model_name = "tacrolimus_HAN2011",
                 model = mm,
                 data = dd,
                 evaluation_type= "sequential_updating",
                 assessment='Bayesian_forecasting' )

res # Print the results
}
}
