% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{update_map_models}
\alias{update_map_models}
\title{Update MAP estimations across occasions for each IDs}
\usage{
update_map_models(
  map_results,
  evaluation_type = c("sequential_updating", "stepwise_updating",
    "sequential_reference_updating", "backward_reference_updating")
)
}
\arguments{
\item{map_results}{A list object containing the MAP estimations obtained in `run_MAP_estimations`
Must include an element named `map_estimations` which stores the posterior estimations
for each occasion.}

\item{evaluation_type}{A character vector specifying the evaluation type to use for updating.
Options include:
- `"sequential_updating"`: Use posterior results sequential_updatingly across all previous occasions.
- `"stepwise_updating"`: Use only the most recent posterior for updating.
- `"sequential_reference_updating"`: Use a chronological reference for posterior updates.
- `"backward_reference_updating"`: Use the most recent chronological reference for updates.
Defaults to `"sequential_updating"`.}
}
\value{
A list containing:
\describe{
  \item{ind_model}{A list of updated individual posterior models for each
  occasion and ID, ready for simulation.}
  \item{eval_type}{The evaluation strategy used.}
}
}
\description{
The `update_map_models` function updates population pharmacokinetic models using MAP posterior estimations
across occasions based on a selected evaluation type. Depending on the selected evaluation
strategy, posterior information is propagated across occasions to generate
individual models ready for simulation. The function returns individual models updated with posterior parameter estimates for each ID at each occasion.
}
\details{
This function evaluates posterior estimations iteratively based on the specified
`evaluation_type`. It ensures compatibility with the evaluation type used during
the creation of the input `map_results`.

The function dynamically names the posterior estimation results, following the pattern
`a.posteriori_occX_Y`, where `X` and `Y` represent the occasions used in the estimation.

The output is intended to be used with [run_pk_simulations()]
}
\examples{
\dontrun{
# Example input data
map_results <- list(
  eval_type = "sequential_updating",
  map_estimations = list(
    "map.estimation.occ_0_1" = NULL,
    "map.estimation.occ_0_1_2" = list() # Replace with real MAP estimation object
  )
)

# Run the function
result <- update_map_models(map_results, evaluation_type = "sequential_updating")
print(result)
}
}
